+++ src/core/hiwire.h
@@ -403,6 +403,9 @@ hiwire_dir(JsRef idobj);
 JsRef
 hiwire_call(JsRef idobj, JsRef idargs);
 
+JsRef
+hiwire_call_bound(JsRef idfunc, JsRef idthis, JsRef idargs);
+
 /**
  * Call a member function.
  *
@@ -441,6 +444,9 @@ hiwire_get_length(JsRef idobj);
 bool
 hiwire_get_bool(JsRef idobj);
 
+bool
+hiwire_is_pyproxy(JsRef idobj);
+
 /**
  * Returns 1 if the object is a function.
  *
@@ -448,6 +454,12 @@ hiwire_get_bool(JsRef idobj);
  */
 bool
 hiwire_is_function(JsRef idobj);
+
+bool
+hiwire_is_error(JsRef idobj);
+
+bool
+hiwire_function_supports_kwargs(JsRef idfunc);
 
 /**
  * Returns true if the object is a promise.